working on tests

+++ src/scripts/tests.sh
@@ -2,7 +2,13 @@
 set -e
 set -x
 
-rm ~/backend-services-api/logging/log.log
+# clear log file if it exists
+if [[ -f "~/backend-services-api/logging/log.log" ]]
+then
+    echo "Removing ~/backend-services-api/logging/log.log"
+    rm ~/backend-services-api/logging/log.log
+fi
+
 echo "log cleared"
 #run pre-commit
 pre-commit run -a
@@ -14,8 +20,15 @@ python3 -m pytest
 sed -i "s/<source>\/home\/mike\/backend-services-api\/src<\/source>/<source>\/github\/workspace\/src<\/source>/g" /home/mike/backend-services-api/src/coverage.xml
 # create coverage-badge
 coverage-badge -o ../coverage.svg -f
-# delete db
-rm ~/backend-services-api/sqlite_db/test.db
+
+
+# remove test.db if it exists
+if [[ -f "~/backend-services-api/sqlite_db/test.db" ]]
+then
+    echo "Removing ~/backend-services-api/sqlite_db/test.db"
+    rm ~/backend-services-api/sqlite_db/test.db
+fi
+
 echo "db removed"
 # generate flake8 report
 flake8 --tee . > flake8_report/report.txt
